Move config, ctrl_msg and handler_common to common/
