+++ controller/handler_ctrl/circuit_request.go
@@ -24,7 +24,7 @@ import (
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/channel/v2"
 	"github.com/openziti/fabric/controller/network"
-	"github.com/openziti/fabric/ctrl_msg"
+	"github.com/openziti/fabric/common/ctrl_msg"
 	"github.com/openziti/fabric/logcontext"
 	"github.com/openziti/fabric/pb/ctrl_pb"
 	"github.com/openziti/identity"
