+++ controller/handler_ctrl/dequiesce_router.go
@@ -20,7 +20,7 @@ import (
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/channel/v2"
 	"github.com/openziti/fabric/controller/network"
-	"github.com/openziti/fabric/handler_common"
+	"github.com/openziti/fabric/common/handler_common"
 	"github.com/openziti/fabric/pb/ctrl_pb"
 )
