+++ controller/handler_ctrl/update_terminator.go
@@ -23,7 +23,7 @@ import (
 	"github.com/openziti/fabric/controller/fields"
 	"github.com/openziti/fabric/controller/network"
 	"github.com/openziti/fabric/controller/xt"
-	"github.com/openziti/fabric/handler_common"
+	"github.com/openziti/fabric/common/handler_common"
 	"github.com/openziti/fabric/pb/ctrl_pb"
 	log "github.com/sirupsen/logrus"
 	"google.golang.org/protobuf/proto"
