+++ controller/network/routesender.go
@@ -24,7 +24,7 @@ import (
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/channel/v2/protobufs"
 	"github.com/openziti/fabric/controller/xt"
-	"github.com/openziti/fabric/ctrl_msg"
+	"github.com/openziti/fabric/common/ctrl_msg"
 	"github.com/openziti/fabric/logcontext"
 	"github.com/openziti/fabric/pb/ctrl_pb"
 	cmap "github.com/orcaman/concurrent-map/v2"
