+++ pb/ctrl_pb/decoder.go
@@ -20,7 +20,7 @@ import (
 	"fmt"
 	"github.com/michaelquigley/pfxlog"
 	"github.com/openziti/channel/v2"
-	"github.com/openziti/fabric/ctrl_msg"
+	"github.com/openziti/fabric/common/ctrl_msg"
 	"github.com/openziti/metrics/metrics_pb"
 	"google.golang.org/protobuf/proto"
 	"strconv"
